changes related to config (see below lines):
Changes:
- added optional config parameter to pyzbar.pyzbar.decode
- the structure is a nested dictionary where the keys are ZBarSymbol objects, and the values are dicts of ZBarConfig-value key-value pairs

Explanation:
I am submitting this PR to accomodate our use case of using I25 barcodes that are only 2 digits. ZBar defaults to a minimum length of 6 for I25 codes, and we had no way to change that in the current pyzbar. While I was in the code, I figured I would make it generic to allow changing any config option(s) for any symbol.

The structure of the config object is admittedly a bit complicated. I decided to make it a nested dict for 2 reasons:
1. to allow changing barcode settings for each individual code rather than for all codes
2. to allow passing only the config object without passing the symbols object